third_party/blink/web_tests/http/tests/security/referrer-policy-attribute-img-no-referrer-when-downgrade-expected.txt

CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/referrer-policy-attribute-img-no-referrer-when-downgrade.html' was loaded over HTTPS, but requested an insecure element 'http://127.0.0.1:8000/security/resources/green-if-no-referrer.php'. This request was not upgraded to HTTPS because its URL's host is an IP address.
